Reduce required crew of korath worldship
Yes I know it makes the worldship pitifully easy to capture now, but who cares if people are able to capture a ship, even if it is vastly better than any other ship in crew. And it definitely does not require 750+ crew just to fly this ship, even if it is the biggest.